Week 3 FacebookClass
Define a new Facebook class which stores a list of FacebookUser objects
A Facebook object is created at the start of the program. The Facebook object is serialized/saved
to "facebook.dat" so the list of users can be reloaded even after closing the program
Also added a couple new custom exceptions IncorrectPasswordException, UsernameAlreadyExistsException
and UsernameNotFoundException which are used now and could be useful in the future maybe.